使用封装的 FMDB 工具类，简化代码

+++ ExTensionDemo/GYTodo.h
@@ -11,8 +11,9 @@
 NS_ASSUME_NONNULL_BEGIN
 
 @interface GYTodo : NSObject
-@property (nonatomic) int id;
+@property (nonatomic) NSInteger id;
 @property (nonatomic, strong) NSString *content;
+
 @end
 
 NS_ASSUME_NONNULL_END
